Added concept of state, nodes now track current state

+++ src/Blockchain/MiningNode.js
@@ -1,6 +1,8 @@
 import Blockchain from "./Blockchain";
 import Transaction from "./Transaction";
 import Worker from "workerize-loader!./MiningNode.worker.js";
+import Wallet from "./Wallet";
+import State from "./State";
 
 class MiningNode {
   constructor({ id, miningNetwork, addBlockToNode }) {
@@ -10,23 +12,39 @@ class MiningNode {
     this.miningWebWorker = new Worker();
     this.miningWebWorker.onmessage = this.handleWorkerMessage.bind(this);
     this.addBlockToNode = addBlockToNode;
+    this.minerWallet = new Wallet();
+    this.state = new State();
+    this.newState = {};
+  }
+
+  async init() {
+    await this.minerWallet.generateKeyPair();
+    this.mine();
   }
 
   mine() {
-    const validTransactions =
+    const initialValidTransactions =
       this.miningNetwork.transactionPool.validTransactions();
-    const rewardTransaction = Transaction.rewardTransaction("mineraddress");
+    const rewardTransaction = Transaction.rewardTransaction(
+      this.minerWallet.publicKey
+    );
+
+    const transactionsToRun = {
+      ...initialValidTransactions,
+      [rewardTransaction.id]: rewardTransaction,
+    };
+
+    const { validTransactions, invalidTransactions, newState } =
+      this.state.runTransactions(transactionsToRun);
+    this.newState = newState;
+
+    this.miningNetwork.transactionPool.removeTransactions(invalidTransactions);
 
     this.miningWebWorker.postMessage({
       action: "MINE",
       lastBlock: this.blockchain.lastBlock(),
       //Web workers are picky about what data types you give them
-      transactions: JSON.parse(
-        JSON.stringify({
-          ...validTransactions,
-          [rewardTransaction.id]: rewardTransaction,
-        })
-      ),
+      transactions: JSON.parse(JSON.stringify(validTransactions)),
     });
   }
 
@@ -36,21 +54,54 @@ class MiningNode {
     if (action === "MINED") {
       this.miningNetwork.broadcastBlock({ block, nodeId: this.id });
       this.blockchain.addBlock(block);
+
       this.addBlockToNode({ nodeId: this.id, newChain: this.blockchain.chain });
       this.miningNetwork.transactionPool.removeTransactions(block.transactions);
+
+      this.state.stateMap = this.newState;
+      this.newState = {};
+
       this.mine();
     }
   }
 
   receiveBlock(block) {
-    this.blockchain.addBlock(block);
+    let validBlock = true;
 
-    this.miningWebWorker.terminate();
-    this.miningWebWorker = new Worker();
-    this.miningWebWorker.onmessage = this.handleWorkerMessage.bind(this);
-    this.addBlockToNode({ nodeId: this.id, newChain: this.blockchain.chain });
+    for (const transactionId in block.transactions) {
+      const transaction = new Transaction(block.transactions[transactionId]);
+      if (!transaction.validTransaction()) {
+        validBlock = false;
+        return;
+      }
+    }
 
-    this.mine();
+    const { _, invalidTransactions, newState } = this.state.runTransactions(
+      block.transactions
+    );
+
+    if (Object.keys(invalidTransactions).length) {
+      validBlock = false;
+      return;
+    }
+
+    if (validBlock) {
+      //Wont be hit if block invalid, but just incase
+      this.blockchain.addBlock(block);
+
+      this.miningWebWorker.terminate();
+      this.miningWebWorker = new Worker();
+      this.miningWebWorker.onmessage = this.handleWorkerMessage.bind(this);
+      this.addBlockToNode({
+        nodeId: this.id,
+        newChain: this.blockchain.chain,
+      });
+
+      this.state.stateMap = newState;
+      this.newState = {};
+
+      this.mine();
+    }
   }
 }
